+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.game.controller.GameController;
+import baseball.game.GameController;
 
 public class Application {
     public static void main(String[] args) {
