--- src/main/java/baseball/domain/Hitter.java
@@ -1,5 +1,0 @@
-package baseball.domain;
-
-public interface Hitter {
-    BallContainer swing(Ball pitchBall);
-}
