--- src/main/java/baseball/domain/Referee.java
@@ -1,5 +1,0 @@
-package baseball.domain;
-
-public interface Referee {
-    Score calcScore(BallContainer ballContainer);
-}
