+++ src/main/java/baseball/domain/object/BallPiece.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.object;
 
 public class BallPiece {
     final int number;
