+++ src/main/java/baseball/domain/object/Hitter.java
@@ -1,0 +1,8 @@
+package baseball.domain.object;
+
+import baseball.domain.object.Ball;
+import baseball.domain.object.BallContainer;
+
+public interface Hitter {
+    BallContainer swing(Ball pitchBall);
+}
