+++ src/main/java/baseball/domain/object/Pitcher.java
@@ -1,4 +1,6 @@
-package baseball.domain;
+package baseball.domain.object;
+
+import baseball.domain.object.Ball;
 
 public interface Pitcher {
     Ball pitch();
