+++ src/main/java/baseball/domain/object/Referee.java
@@ -1,0 +1,8 @@
+package baseball.domain.object;
+
+import baseball.domain.object.BallContainer;
+import baseball.domain.score.Score;
+
+public interface Referee {
+    Score calcScore(BallContainer ballContainer);
+}
