+++ src/main/java/baseball/domain/service/RandomBallGenerator.java
@@ -1,6 +1,6 @@
 package baseball.domain.service;
 
-import baseball.domain.Ball;
+import baseball.domain.object.Ball;
 import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.HashSet;
