+++ src/main/java/baseball/game/stage/ScoreCalcStage.java
@@ -1,8 +1,8 @@
 package baseball.game.stage;
 
 import baseball.constant.Constants;
-import baseball.domain.Score;
-import baseball.game.controller.GameController;
+import baseball.domain.score.Score;
+import baseball.game.GameController;
 import baseball.game.domain.ObserveRequest;
 import baseball.game.enums.GameState;
 import baseball.game.domain.GameConfig;
